backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
supabase==2.27.2
openai==1.59.8
python-dotenv==1.2.1
PyJWT==2.10.1
pydantic==2.12.5
